+++ rypper-utils/src/fetch/mod.rs
@@ -1,3 +1,0 @@
-pub mod metadata;
-pub mod mirrorlist;
-pub mod repofiles;